Updated Navigation drawer layout
