+++ .idea/misc.xml
@@ -32,7 +32,7 @@
         <entry key="app/src/main/res/layout/fragment_item_list_dialog_list_dialog.xml" value="0.14235055724417425" />
         <entry key="app/src/main/res/layout/fragment_item_list_dialog_list_dialog_item.xml" value="0.14235055724417425" />
         <entry key="app/src/main/res/layout/fragment_login.xml" value="0.19022327367810235" />
-        <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.5158132530120482" />
+        <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.23310909041439193" />
         <entry key="app/src/main/res/layout/fragment_recipe_details.xml" value="0.2011664717726221" />
         <entry key="app/src/main/res/layout/fragment_recipe_search.xml" value="0.20568927789934355" />
         <entry key="app/src/main/res/layout/fragment_register.xml" value="0.41893115942028986" />
