+++ app/src/main/res/layout/fragment_profile.xml
@@ -169,5 +169,22 @@
         app:layout_constraintTop_toTopOf="parent"
         app:menu="@menu/menu_navigation_drawer">
 
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom"
+            android:orientation="vertical"
+            android:padding="16dp"
+            android:paddingStart="30dp">
+
+            <TextView
+                android:id="@+id/logout"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:drawableStart="@drawable/ic_baseline_logout_24"
+                android:gravity="center"
+                android:text="Log out" />
+        </LinearLayout>
+
     </com.google.android.material.navigation.NavigationView>
 </androidx.constraintlayout.widget.ConstraintLayout>
